pd: ♎ move node entrypoint into library
fixes #3695.

this moves the startup code into a place where we'll be able to exercise
it in tests someday. no changes are made to the contents of this
expression.
